bridge/python-csv/requirements.txt
flask==2.3.2
requests==2.26.0
psycopg2-binary==2.9.6
flask-cors==3.0.10

